Upgrade C++ to protobuf 3.12
This has various performance improvements, and also enables Arena
support by default (used in a follow-on commit)

Reviewed-on: http://gerrit.cloudera.org:8080/16135
Tested-by: Kudu Jenkins

+++ src/kudu/consensus/log.h
@@ -615,7 +615,7 @@ class LogEntryBatch {
   LogEntryBatchPB entry_batch_pb_;
 
    // Total size in bytes of all entries
-  const uint32_t total_size_bytes_;
+  const size_t total_size_bytes_;
 
   // Number of entries in 'entry_batch_pb_'
   const size_t count_;
